Refactory code for users, videogames and comments

+++ src/main/java/antonioloiacono/tesi/monolitica/repository/CommentRepository.java
@@ -3,4 +3,4 @@ package antonioloiacono.tesi.monolitica.repository;
 import antonioloiacono.tesi.monolitica.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {}
+public interface CommentRepository extends JpaRepository<Comment, Long> {}
